refactor: 피드 본인 확인 로직 목적에 맞게 auth 도메인으로 이동 (#181)
