+++ src/entities/user/index.ts
@@ -1,4 +1,3 @@
 export * from './api/user'
-export * from './config/bioQuotes'
 export * from './model/useUser'
 export * from './types/user'
