+++ src/entities/user/types/user.ts
@@ -1,9 +1,5 @@
 import type { UserInfo } from '@/features/auth'
 
-export interface ShareCodeOwnerResponse {
-  isOwner: boolean
-}
-
 // TODO: follow 브랜치 병합되면 해당 type extends
 export interface ProfileResponse extends UserInfo {
   bio: string
