+++ src/features/auth/model/useAuth.ts
@@ -5,7 +5,9 @@ import {
   getAnonymousLogin,
   getUserInfo,
   deleteAccount,
+  getShareCodeOwner,
   type LoginPayload,
+  type ShareCode,
 } from '@/features/auth'
 
 export const useLogin = () => {
@@ -33,5 +35,21 @@ export const useDeleteAccount = () => {
   return useMutation({
     mutationKey: ['deleteAccount'],
     mutationFn: deleteAccount,
+  })
+}
+
+export const useVerifyOwner = () => {
+  return useMutation({
+    mutationKey: ['verifyOwner'],
+    mutationFn: (shareCode: ShareCode) => getShareCodeOwner(shareCode),
+  })
+}
+
+export const useOwnerStatus = (shareCode: ShareCode) => {
+  return useQuery({
+    queryKey: ['ownerStatus', shareCode],
+    queryFn: () => getShareCodeOwner(shareCode),
+    enabled: !!shareCode,
+    staleTime: 1000 * 60 * 5, // 5분
   })
 }
