+++ src/pages/feed/index.tsx
@@ -6,7 +6,8 @@ import styled, { css } from 'styled-components'
 import { useToast } from '@/app/providers'
 import { Gear, Pencil, Add, Share } from '@/assets/icons'
 import { FeedBg } from '@/assets/images'
-import { useOwnerStatus, useUserProfile, getRandomBio } from '@/entities/user'
+import { useUserProfile } from '@/entities/user'
+import { useOwnerStatus } from '@/features/auth'
 import { FeedbackIcon } from '@/pages/feedback/ui'
 import { useCopyShareUrl } from '@/shared/lib'
 import { flexRowCenter } from '@/shared/styles/mixins'
@@ -31,12 +32,21 @@ const FeedPage = () => {
     isError: isProfileError,
   } = useUserProfile(shareCode)
 
+  const RANDOM_BIO_QUOTES = [
+    '재생 목록에\n어떤 곡이 있나요?',
+    '오늘의 기분에는\n이 트랙이!',
+    '짧은 글로 나의\n음악 취향을 알려줘!',
+    '나를 자유롭게\n소개해줘!',
+  ] as const
+
   const currentTab = searchParams?.get('tab') || ('cds' as FEED_TAB_TYPE)
   console.log('currentTab', currentTab)
   console.log('userProfile', userProfile)
 
   const bioQuotes = useMemo(() => {
-    return userProfile?.bio || getRandomBio()
+    if (userProfile?.bio) return userProfile?.bio
+    const randomIndex = Math.floor(Math.random() * RANDOM_BIO_QUOTES.length)
+    return RANDOM_BIO_QUOTES[randomIndex]
   }, [userProfile?.bio])
 
   const onShareButtonClick = async () => {
